Rename CrmUser_user component file
